+++ pymince/__init__.py
@@ -1,3 +1,3 @@
 # -*- coding: utf-8 -*-
 __title__ = "pymince"
-__version__ = "2.2.1"
+__version__ = "2.3.0"