+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## In Progress ##
 
+## v2.3.0 (2023-01-13) ##
+
 - **Added:** iterator.mul, iterator.truediv, iterator.sub
 - **Added:** functional.set_attributes
 - **Added:** json.JSONEncoder. Handles additional types compared to `json.JSONEncoder`